Update SpecialPower.ini
Scud Storm revert to 5 mins:
Nerf was unnecesarry. The main approach of 1.04p is to buff the weaknesses. A nerf is only used in cases where an unfair advantage is present relative to other factions (Air Force chinook e.g.). With the BT included the balance is now like so:
Particle Cannon 4+1 minutes, Nuclear Missile 5+1 minutes, Scud Storm 5+2 minutes.